Implement search
Including the following changes:

- Synchronised calling of the methods which initiate HTTP-GETs of the receipts
- A lock for the synchronisation
- Mutex-like boolean class-variables which are only released inside of the call-back of the GET, and correspondingly only flagged to true inside the synchronised block
- Pull-to-refresh
- GUI-bugfixes

Note that as the synchronised methods only alter the boolean variables to a true state inside of the synchronised blocks, this does in effect allow us to defer from initiating additional fetches until success or failure of the GET-call. In other words the lock lets us build up a queue-like structure for asynchronous back-end-calls, which do not alter the internal data-structures simultaneously, nor result in deadlocks (as only one lock is employed). Conversely, blocking or synchronous fetching would make the GUI freeze until the result returns, as the VC-thread calls the APIClient-fetch.